CRUD Categoria

+++ Models/DelCategoriaPHP.php
@@ -1,6 +1,13 @@
 <?php
 include("../Controller/Config.php");
 $cursoID = $_GET['del'];
-mysqli_query($mysqli,"DELETE FROM TipoCurso where AT_TipoCursoID = $cursoID");
+if (mysqli_query($mysqli,"DELETE FROM TipoCurso where AT_TipoCursoID = $cursoID")) {
+  mysqli_query($mysqli,"DELETE FROM TipoCurso where AT_TipoCursoID = $cursoID");
+  header("Location: ../Views/Cursos.php?status=5");
+}else{
+  header("Location: ../Views/Cursos.php?status=6");
+
+}
+
 
  ?>
